C3. Making changes in some variable types

+++ src/main/java/com/milnesium/Vehicle.java
@@ -13,7 +13,7 @@ public class Vehicle {
     //Additional
     //Adding a boost option to the car, it can get the driver in the 1st place but might decrease much faster the fuel
     boolean boostSpeed;
-    byte boostConsumptionFactor;
+    float boostConsumptionFactor;
     float boostSpeedFactor;
     int boostLevel;
